requirements.txt
click>=8.1.3
kafka-python>=2.0.2
PyYAML>=6.0
python-dotenv>=1.0.0
